+added custom routes and fixes
