+++ src/pages/404.tsx
@@ -1,0 +1,6 @@
+import { ReactElement } from "react";
+
+export default function ErrorPage(): ReactElement {
+  return <>Hey!{404}</>;
+}
+
